주문 금액 타입 long으로 변경

+++ src/main/java/com/programmers/be14/nbe12141team03/domain/order/dto/mergedShipment/MergedShipmentResponse.java
@@ -10,6 +10,6 @@ public record MergedShipmentResponse(
         String zipCode,
         int orderCount,
         List<Long> mergedOrderIds,
-        int totalPrice,
+        long totalPrice,
         List<MergedItemResponse> orderItemList
 ) {}
